backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
torch==2.2.2
numpy
scipy
scikit-learn>=1.3.0
pytest==8.3.4
httpx==0.28.1
google-generativeai>=0.3.0
